Update the n-gram crawler to use lookahead
Basically this means that when choosing the next page, we look ahead x
number of pages down the link tree to maximize our potential score.
Currently we're using lookahead 1, which is the same behavior as before
this change.

Also, we now use the `ScoredLink` type rather than tuples of link and
score.