Score a page as infinity if it is the destination page
I've never actually thoroughly dug into these before now. We have an
issue that the n-gram models aren't normalized for the length of the
text (unsure how to do that yet), so in fact the destination page can
often be scored lower than other pages if it is long. That needs to be
fixed separately.